// src/app/interfaces/cotizacion/cliente.ts
export interface Cliente {
  cliec_icod_cliente: number;
  cliec_vnombre_cliente: string;
  cliec_vapellido_paterno: string;
  cliec_vapellido_materno: string;
  cliec_vdireccion_cliente: string;
  cliec_vnro_telefono: string;
  cliec_cruc: string;
  cliec_observacion1: string;
  cliec_observacion2: string;
  situacion: string;
  tipoDoc: string;
  cliec_vnumero_doc_cli:string;
  cliec_vnombre_contacto:string;
  cliec_vcod_cliente:number;
}