// src/app/interfaces/cotizacion/ProdTerminado.ts
export interface ProductoTerminado {
  prdc_icod_producto: number;
  prdc_vcode_producto: string;
  prdc_vdescripcion_larga: string;
  prdc_isituacion: boolean;
  strUnidadMedida: string;
  prdc_vcaracteristica: string;
  prodc_icod_tipo: number;
  prodc_nomenclatura: string;
  prodc_nom_lab: string;
  prodc_p_a: string;
  prodc_env_asociado: string;
  prodc_vsku: string;
  prodc_desc_tipo_unidad_medida: string;
  desCate: string;
  prodc_desc_tipo_tabla: string;
  prod_nvalor_costo: number;
  prod_nprecio_venta_vendedor: number;
  prod_nprecio_venta_publico: number;
  formaFar: string;
  prdc_des_catalogo: string;
}